requirements.txt
Flask
Flask-MySQL-Connector